Updated snappy cffi to use CFFI 1.0 out-of-line API mode
